add: full common and admin hadnler

+++ internal/storage/sql/application.go
@@ -2,7 +2,7 @@ package sql
 
 import (
 	"context"
-	"errors"
+	"studentRecordsApp/internal/service"
 
 	"github.com/google/uuid"
 	"github.com/jmoiron/sqlx"
@@ -12,12 +12,14 @@ import (
 	"studentRecordsApp/internal/storage/sql/sqlEntities"
 )
 
+var _ service.ApplicationDb = (*Application)(nil)
+
 type Application struct {
 	db *sqlx.DB
 }
 
-func NewApplication(db *sqlx.DB) Application {
-	return Application{
+func NewApplication(db *sqlx.DB) *Application {
+	return &Application{
 		db: db,
 	}
 }
@@ -37,7 +39,7 @@ func (a *Application) Get(ctx context.Context) ([]entities.Application, error) {
 	return results, nil
 }
 
-func (a *Application) GetByUserId(ctx context.Context, userId uuid.UUID) ([]entities.Application, error) {
+func (a *Application) GetForUser(ctx context.Context, userId uuid.UUID) ([]entities.Application, error) {
 	var result []sqlEntities.Application
 	err := a.db.SelectContext(ctx, &result, "SELECT * FROM Applications WHERE student_id = $1 ORDER BY created_at",
 		userId)
@@ -64,14 +66,20 @@ func (a *Application) GetById(ctx context.Context, id, userID uuid.UUID) (entiti
 	return casts.ApplicationSqlToService(ctx, result), nil
 }
 
-func (a *Application) Add(ctx context.Context, app entities.Application) error {
-	tx, isOk := ctx.Value("tx").(*sqlx.Tx)
-	if !isOk {
-		return errors.New("tx is not set")
+func (a *Application) GetOne(ctx context.Context, id uuid.UUID) (entities.Application, error) {
+	var result sqlEntities.Application
+	err := a.db.GetContext(ctx, &result, "SELECT * FROM Applications WHERE id = $1",
+		id)
+	if err != nil {
+		return entities.Application{}, err
 	}
 
-	_, err := (*tx).ExecContext(ctx, `INSERT INTO Applications (id, student_id, application_name, 
-                          contact_info, application_text, application_status, created_at, link_to_application) 
+	return casts.ApplicationSqlToService(ctx, result), nil
+}
+
+func (a *Application) Add(ctx context.Context, app entities.Application) error {
+	_, err := a.db.ExecContext(ctx, `INSERT INTO Applications (id, student_id, _name, 
+                          contact_info, _text, status, created_at, link) 
     										VALUES ($1, $2, $3, $4, $5, $6, $7, $8)`,
 		app.Id, app.StudentId, app.Name, app.ContactInfo, app.Text, app.Status, app.CreatedAt, app.Link)
 
@@ -79,21 +87,21 @@ func (a *Application) Add(ctx context.Context, app entities.Application) error {
 }
 
 func (a *Application) Update(ctx context.Context, app entities.Application) error {
-	_, err := a.db.ExecContext(ctx, "UPDATE Applications SET application_name = $1, contact_info = $2, application_text = $3 WHERE id = $4",
+	_, err := a.db.ExecContext(ctx, "UPDATE Applications SET _name = $1, contact_info = $2, _text = $3 WHERE id = $4",
 		app.Name, app.ContactInfo, app.Text, app.Id)
 
 	return err
 }
 
-func (a *Application) SetClose(ctx context.Context, id uuid.UUID) error {
-	_, err := a.db.ExecContext(ctx, `UPDATE Applications SET application_status = $1 WHERE id = $2`,
-		entities.ApplicationClosed, id)
+func (a *Application) UpdateStatus(ctx context.Context, id uuid.UUID, status string) error {
+	_, err := a.db.ExecContext(ctx, `UPDATE Applications SET status = $1 WHERE id = $2`,
+		status, id)
 
 	return err
 }
 
-func (a *Application) Delete(ctx context.Context, id uuid.UUID) error {
-	_, err := a.db.ExecContext(ctx, "DELETE FROM Applications WHERE id = $1", id)
+func (a *Application) Delete(ctx context.Context, id, userId uuid.UUID) error {
+	_, err := a.db.ExecContext(ctx, "DELETE FROM Applications WHERE id = $1 AND student_id = $2", id, userId)
 
 	return err
 }
